+++ app/src/main/java/com/shining/webhandler/common/data/WebData.kt
@@ -3,10 +3,10 @@ package com.shining.webhandler.common.data
 import android.graphics.Bitmap
 
 /**
- * FavoriteData.kt
+ * WebData.kt
  * WebHandler
  */
-data class WebData(val id: UInt,
+data class WebData(val id: Long,
                    var icon: Bitmap? = null,
                    var title: String = "",
                    val url: String,
